+++ crates/block-producer/src/test_utils/block.rs
@@ -22,7 +22,14 @@ pub async fn build_expected_block_header(
     store: &MockStoreSuccess,
     batches: &[TransactionBatch],
 ) -> BlockHeader {
-    let last_block_header = *store.last_block_header.read().await;
+    let last_block_header = *store
+        .block_headers
+        .read()
+        .await
+        .iter()
+        .max_by_key(|(block_num, _)| *block_num)
+        .unwrap()
+        .1;
 
     // Compute new account root
     let updated_accounts: Vec<_> =
@@ -104,7 +111,14 @@ impl MockBlockBuilder {
         Self {
             store_accounts: store.accounts.read().await.clone(),
             store_chain_mmr: store.chain_mmr.read().await.clone(),
-            last_block_header: *store.last_block_header.read().await,
+            last_block_header: *store
+                .block_headers
+                .read()
+                .await
+                .iter()
+                .max_by_key(|(block_num, _)| *block_num)
+                .unwrap()
+                .1,
 
             updated_accounts: None,
             created_notes: None,